#324: Switch pad_value type to float and add casting to tensor dtype

+++ libs/tensor/tensor.cpp
@@ -199,7 +199,7 @@ void Tensor::print(Layout print_layout, bool pretty_print) const {
     tensor_impl::print_wrapper(*this, print_layout, pretty_print);
 }
 
-Tensor Tensor::pad(const std::array<uint32_t, 4> &output_tensor_shape, const std::array<uint32_t, 4> &input_tensor_start, uint32_t pad_value) const {
+Tensor Tensor::pad(const std::array<uint32_t, 4> &output_tensor_shape, const std::array<uint32_t, 4> &input_tensor_start, float pad_value) const {
     TT_ASSERT(on_host() && "Tensor must be on host for padding");
     TT_ASSERT(this->layout() == Layout::ROW_MAJOR && "Tensor layout must be ROW_MAJOR for padding");
     return tensor_impl::pad_wrapper(*this, output_tensor_shape, input_tensor_start, pad_value);
